Initial port to 1.21

+++ src/tools/java/mekanism/tools/common/item/ItemMekanismSword.java
@@ -13,7 +13,8 @@ import org.jetbrains.annotations.NotNull;
 public class ItemMekanismSword extends SwordItem {
 
     public ItemMekanismSword(MaterialCreator material, Item.Properties properties) {
-        super(material, properties.attributes(createAttributes(material, material.getSwordDamage(), material.getSwordAtkSpeed())));
+        //TODO - 1.21: Ensure the patch to neo that adds a float variant gets added back in
+        super(material, properties.attributes(createAttributes(material, (int) material.getSwordDamage(), material.getSwordAtkSpeed())));
     }
 
     @Override
